+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## NEXT
 
 * Adds `executeTransation`
+* Adds `putSingle` and `updateSingle` in `Queryflow`
 
 ## 0.0.1